Introduce HomeParser
